backend/app/auth.py
from fastapi import APIRouter, Depends, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from pydantic import BaseModel
from typing import Optional
import jwt
import requests
import os
from datetime import datetime, timedelta
import logging

router = APIRouter()
security = HTTPBearer()
logger = logging.getLogger(__name__)

# Pydantic models
class UserInfo(BaseModel):
    user_id: str
    email: str
    role: str = "user"

class LoginRequest(BaseModel):
    email: str
    password: str

class TokenResponse(BaseModel):
    access_token: str
    token_type: str
    expires_in: int

# AWS Cognito configuration
COGNITO_USER_POOL_ID = os.getenv("COGNITO_USER_POOL_ID")
COGNITO_CLIENT_ID = os.getenv("COGNITO_CLIENT_ID")
COGNITO_REGION = os.getenv("COGNITO_REGION", "us-east-1")

# JWT validation function
async def verify_token(credentials: HTTPAuthorizationCredentials = Depends(security)) -> UserInfo:
    """
    Verify JWT token from AWS Cognito
    """
    token = credentials.credentials
    
    try:
        # For development - simplified validation
        # In production, implement proper Cognito JWT validation
        if token == "dev-token":
            return UserInfo(
                user_id="dev-user",
                email="dev@example.com",
                role="admin"
            )
        
        # TODO: Implement proper Cognito JWT validation
        # 1. Download Cognito public keys
        # 2. Verify token signature
        # 3. Validate claims (iss, aud, exp, etc.)
        
        # Placeholder validation
        decoded = jwt.decode(token, options={"verify_signature": False})
        
        return UserInfo(
            user_id=decoded.get("sub", "unknown"),
            email=decoded.get("email", "unknown@example.com"),
            role=decoded.get("custom:role", "user")
        )
        
    except jwt.InvalidTokenError as e:
        logger.error(f"Invalid token: {e}")
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid authentication token",
            headers={"WWW-Authenticate": "Bearer"},
        )
    except Exception as e:
        logger.error(f"Token validation error: {e}")
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Authentication failed",
            headers={"WWW-Authenticate": "Bearer"},
        )

@router.post("/login", response_model=TokenResponse)
async def login(request: LoginRequest):
    """
    Authenticate user with AWS Cognito
    """
    try:
        # For development - return mock token
        if request.email == "dev@example.com" and request.password == "password":
            return TokenResponse(
                access_token="dev-token",
                token_type="bearer",
                expires_in=3600
            )
        
        # TODO: Implement actual Cognito authentication
        # Use boto3 to authenticate with Cognito User Pool
        
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid credentials"
        )
        
    except Exception as e:
        logger.error(f"Login error: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="Authentication service error"
        )

@router.get("/me", response_model=UserInfo)
async def get_current_user(current_user: UserInfo = Depends(verify_token)):
    """
    Get current authenticated user information
    """
    return current_user

@router.post("/refresh")
async def refresh_token(credentials: HTTPAuthorizationCredentials = Depends(security)):
    """
    Refresh JWT token
    """
    # TODO: Implement token refresh logic
    return {"message": "Token refresh not implemented yet"}

@router.post("/logout")
async def logout(current_user: UserInfo = Depends(verify_token)):
    """
    Logout user (invalidate token)
    """
    # TODO: Implement logout logic (token blacklisting)
    return {"message": "Logged out successfully"}