backend/requirements.txt
# Python 3.13.7 compatible requirements
# For faster installs, use: uv pip install -r requirements.txt
# 
# Updated for Python 3.13 compatibility - some packages use latest versions

# Core FastAPI and server
fastapi[all]==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# AI and ML dependencies (Python 3.13 compatible - simplified approach)
langchain==0.3.1
langchain-community==0.3.1
langchain-openai==0.2.1
pymupdf==1.24.10
# Using simple approach for Python 3.13 compatibility until full ML ecosystem catches up
# chromadb==0.4.24  # Uncomment when Python 3.13 wheels are available
# faiss-cpu==1.8.0  # Uncomment when Python 3.13 wheels are available
openai==1.51.0

# AWS dependencies
boto3==1.35.36

# Database dependencies
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
alembic==1.13.3

# Background tasks and caching
celery==5.4.0
redis==5.1.1

# WebSocket and real-time communication
python-socketio==5.11.4
websockets==13.1

# Utilities
requests==2.32.3
aiofiles==24.1.0
pillow==10.4.0
pydantic==2.9.2